Wrap some sfdx commands, add integration tests.
Decided to wrap up some of the commands the SFDX Orb provides in some
custom commands in this Orb so we can build upon them later without
breaking the simplicity/flexibility of the existing SFDX Orb.

This includes scratch-up (creates and opens scratch), scratch-down (only
deletes the org for now, but we can expand this later, e.g removing
permissions?), and setup, which installs the CLI and authenticates at
once. There is some potential for additional steps to be done here
specific to APEX application scratch orgs.